+++ kernels/common/scene_triangle_mesh.cpp
@@ -39,8 +39,8 @@ namespace embree
   
   void TriangleMesh::disabling() 
   { 
-    if (numTimeSteps == 1) { atomic_add(&parent->numTriangles ,-numTriangles); }
-    else                   { atomic_add(&parent->numTriangles2,-numTriangles); }
+    if (numTimeSteps == 1) { atomic_add(&parent->numTriangles ,-(ssize_t)numTriangles); }
+	else                   { atomic_add(&parent->numTriangles2,-(ssize_t)numTriangles); }
   }
 
   void TriangleMesh::setMask (unsigned mask)
